+++ changelog.md
@@ -1,10 +1,14 @@
 # git-off-my-land changelog
 
+## v1.1.0
+* Change: require -> import
+* Add: unit tests for filterFilesList()
+* Fix: Linter errors
+
 ## v1.0.1
 * Remove: unnecessary comments from lib
 * Change: License to Apache 2, from MIT
 * Remove: NPM run script
-
 
 ## v1.0.0
 Initial version